math-first-principles-tutor/requirements.txt
fastapi
uvicorn
streamlit
requests
smolagents
litellm
langfuse
sympy
matplotlib
numpy
pillow
python-dotenv